MAS: deterministic METIS seed, sanity check for mesh_part, contact-aware benchmark fix
- mesh_partition: use fixed METIS seed (0) for deterministic partitioning
- mesh_partition: keep output validation (non-negative IDs, size limit, contiguous)
- sanity_check: add MeshPartitionCheck for FEM geometries
  - Error: negative IDs, empty partitions (gaps), size > 32
  - Info: size < 4, not aligned to 4/8/16/32
- contact-aware benchmark: use 80x80 self-colliding cloth, 5% tolerance
- scene_default_config: fix include path (_config.h -> scene_default_config.h)

+++ apps/tests/sim_case/62_fem_mas_contact_aware_benchmark.cpp
@@ -184,6 +184,7 @@ TEST_CASE("62_fem_mas_contact_aware_benchmark", "[fem][mas][benchmark]")
                      static_cast<double>(off_spmv_count) / on_spmv_count);
     }
 
-    // Contact-aware should not significantly increase SpMV count.
-    REQUIRE(on_spmv_count <= off_spmv_count * 1.05);
+    // Contact-aware should not significantly regress.
+    // Allow 5% tolerance since hierarchy changes can shift Newton convergence slightly.
+    REQUIRE(on_spmv_count <= static_cast<int>(off_spmv_count * 1.05));
 }
